+++ src/Input/input.hxx
@@ -191,7 +191,7 @@ public:
   virtual void init ();
   virtual void bind ();
   virtual void unbind ();
-  virtual void update (int dt);
+  virtual void update (double dt);
 
 
   /**
