+++ src/Main/logger.hxx
@@ -44,7 +44,7 @@ public:
   virtual void init ();
   virtual void bind ();
   virtual void unbind ();
-  virtual void update (int dt);
+  virtual void update (double dt);
 
 private:
