+++ src/Main/viewmgr.hxx
@@ -59,7 +59,7 @@ public:
     virtual void init ();
     virtual void bind ();
     virtual void unbind ();
-    virtual void update (int dt);
+    virtual void update (double dt);
 
     // getters
     inline int size() const { return views.size(); }
